requirements.txt
python-dotenv
langchain
openai
google-api-python-client # Google API
google-search-results # SerpAPI
youtube-transcript-api
google-search-results
chromadb
tiktoken
faiss-cpu
wikipedia
cohere
sendgrid
pywhatkit
streamlit_chat
wolframalpha
pyowm
chainlit
git+https://github.com/openai/whisper.git
judini
duckduckgo-search
arxiv